Add option for a TeX Live profile (fixes #6)
A profile can be specified with the --profile option
of install-tl-ubuntu. This is passed on to install-tl.
An example.profile is included with the repository to
give users a starting point for creating a profile.

A few of the checks are disabled if a profile is used because
install-tl-ubuntu is not smart enough to adapt depending on
which components are selected for installation.  For example,
font configuration commands do not report errors.