Make script runnable as a stand-alone (fixes #3)
The file debian-control-texlive-in will now be downloaded
if it is not in the directory from which install-tl-ubuntu is run.
It is the only dependency and thus now only install-tl-ubuntu is
needed for things to work. I've learned that many people use Github
by downloading files individually. This commit makes it so they only
need to download one file.

An alternative would be to embed the control file in the main script,
but this would remove the flexibility of modifying the control file
while updating to the latest install-tl-ubuntu. One such use case is
that one might want to modify the Debian control file so that Ubuntu
installs a certain (older) version of a package. This should then
probably (unless you want both versions) be accompanied by a change
in the TeX Live profile (use --profile and see example.profile).