Add script to find equivs and improve control (#7)
The new script findEquivs has the goal of finding Debian packages
that a manual installation of TeX Live replaces. It works as follows:

(1) use apt-file to find packages that contain a file in /opt/texbin
(2) see which packages from (1) have all their binaries contained
    in /opt/texbin

This script could be improved to reduce false positives and to detect
packages that are equivalent that do not contain binaries. See
comments in the script for more ideas to improve it.

The following packages (found by findEquivs) are added:

  texlive-lang-cjk
  asymptote
  biber
  jadetex
  makejvf
  mendexk
  m-tx
  pdfjam
  pmx
  xmltex