Overwrite existing /opt/texbin link (fixes #11)
This arises if there was a previous TeX Live installation,
such as if installing TeX Live 2014 when Tex Live 2013
was previously installed and not manually removed.

A warning is given to the user in case he/she would like to
manually remove the old TeX Live installation.

Thanks to adin for reporting the issue.