Fix installation of culmus* packages for --hebrew
An incorrect usage of double quotes caused "culmus culmus-fancy" to
be treated as one package.

The mistake was introduced at 2210d430, but the problem was not
binding until 130f01a7 when the variable HEBREW_PACKS was used to
store the names of more than one package.

Thanks to izar for reporting the problem and sending the logs, which
made it easy to figure out the cause.