Remove an unneeded workaround for the AA package
The downloaded aa-package.tar.gz thankfully no longer bundles an old
version of natbib.sty, which could have caused that version to be
used instead of the TeX distribution's natbib.sty. It seems enough
users must have complained.

I have not checked whether the other AA workound (which follows
immediately the code that this commit removes) is still necessary,
but leaving that workaround is harmless.