Fix the repository arg that's passed to install-tl
I'm not sure if the paths on the mirror webpage changed or the way
in which the -repository argument to install-tl must be specified
was changed. In either case, the -repository argument now works when
installing TeX Live 2016.

Note that no change was necessary for TeX Live 2016 pretest so this
change will not affect using the --pretest option.