Only install 'parallel' package when necessary
It is not necessary when the --iso or --repository option is used.

It is possible that this could fix the issue at #30. The reason is
that if 'parallel' fails to install, the variable TLREPO is set to a
default value. So in the code if there is a check that assumes that
a non-empty TLREPO variable means an ISO is not being used, then
this would explain #30.

However, I'm guessing there are additional problems in #30.