--more-tex: detect & rm bundling of external files
The detection works as follows: the third-party LaTeX packages
directory is searched after downloading and before running texhash,
and then kpsewhich is run on the .cls, .sty, and .bst files. If
kpsewhich finds a match for a file, it means that the filename
exists already from the TL installation. In that case, the file is
removed from the third-party package.

This "bundle-removal" feature will be used in the next commit, when
support is added for Copernicus (which currently bundles some files
that have TL versions).