--more-tex: add old AGU-LaTeX files
These American Geophysical Union files are old and the newer files
should be preferred for new documents. See here:

  https://www.agu.org/-/media/Files/Publications/Latex_submission_guidelines_Sept52019.pdf

The newer files, e.g., "agujournal2019.cls", include the year in the
filename so do not conflict with the older "agutex.cls" that this
commit adds to the archive. These old files are in the archive to
preserve compilation of some older documents of users.