Install newer TL version in parallel
If previous installation done by install-tl-ubuntu is detected,
on the current run set things up for the new installation to work,
but such that the old installation can still be used by using
"switch-tl-to". To do this, the user needs to source the "switches"
file from their .bashrc.

This feature works also for the special case where the user is
installing TL 2023 with an existing TL 2023. e.g., maybe the user
wanted installation of the same TL with different package profiles.