Use more informative name for TL dir
Instead of, e.g., "2024", use "TL24_2024-05-20" to include the date
on which that directory was created.

This directory naming is useful for backup schemes. For example, a
user script may backup that directory before doing a tlmgr update,
and then could choose to rename the directory to reflect the new
status after the tlmgr update.